fits: ignore ZQUANTIZ keyword
Sogo Mineo has pointed out that afw can write a ZQUANTIZ header
keyword (e.g., it is picked up when we read a compressed FITS
header we wrote, and then that header is transferred to a new
file), which can break ds9's handling of mask extensions.

ZQUANTIZ is related to how FITS tile compression is done, and
so we should never read or write it (we leave it up to cfitsio).