Simplify BackgroundList.readFits to avoid file leaks.
Instead of catching the FitsError raised when trying to read
a non-existent HDU, we just ask for the number of HDUs up
front and then iterate.  This seems to be enough to avoid the file
leak.

This change also modified the actual image and mask reading to use
the new Reader classes, mostly for readability; the constructors used
previously trivially forward to the Readers in C++, so it seems
unlikely that change is playing a part in avoiding the leak.