Fix bad stringification in Separable ellipses.
We'd been using %-style format strings with pybind11's wrapper for
string.format(), which is surprisingly a silent failure, not a loud
one.

Other ellipse classes did not suffer from this problem because their
__str__ and __repr__ implementations were defined in Python.

There's some pretty obvious (preexisting) code duplication here, but
fixing that would be a much bigger change than I want to make as an
aside on this ticket.