Rewrite ellipses::PixelRegion as a container.
The previous implementation as a lazy iterator was always a bit
of a premature optimization (which may still be generous - it may
have just been unnecessary cleverness).

What's worse, it was slightly but critically wrong, because it
declared that iterator to be a RandomAccessIterator while only meeting
the requirements of InputIterator.

However, even after fixing that, we get still get wonky iterator
behavior in Python!  It looks like pybind11::make_iterator is just
assuming all wrapped iterators satisfy at least ForwardIterator.
Using a real ForwardIterator (actually a RandomAccessIterator) by
removing the laziness and returning a std::vector::const_iterator
does fix that.