cameraGeom: add applyExternalPhotoCalibCallback
This adds a callback function to afw.cameraGeom.utils to apply an
external photometric calibration to the ccd image read in with
getCcdImage() within the BulterImage class.  This also adds the
(optional) parameter asMaskedImage to makeImageFromCamera() so
that it can be passed through to the getCcdImage() function from
a call to it and thus has the ability to get a maskedImage returned.