ExposureFitsReader should not read the mask/variance if they are missing
If a compressed single image FITS file is read an attempt is made
to read the mask and variance because additional extensions
can be present.   This will fail because they aren't really
masks but on failing an assumption is made that the file
was registered with the reader. This is not set because
nextHdu can return a null pointer.  Therefore add
protections such that we do not attempt to read the
mask/variance if the fits file attribute is null.