Redocument Amplifier getters to reflect actual usage.
Actual usage is indeed inconsistent with intended usage, but the
original design didn't have an answer for how to represent amplifier
geometry for images that weren't yet assembled, so it's not surprising
that we haven't been using it as designed.  In any case, it's long past
time to reflect this in the documentation.