Allow parent to be an iterable
The method `getChildren` ensures that the afw catalog is sorted by the
parentKey before fetching the children for a given parentId. Checking
for pre-condition is more expensive than the operation of fetching
children. The new interface allows passing an iterable of parent Ids
so that the pre-condition check happens only once for all the parents.