Allow writeFitsWithOptions for image/mask to select the options section
Image and Mask previously always used an "image" section of the
compression options. This made it impossible to reuse the compression
settings from an Exposure without copying the relevant section to
image. This change allows the section to be explicitly specified
so that a user can declare that they want the variance settings
to be applied. Mask is changed to always prefer "mask" section
before falling back to "image".