Move Catalog's caching of its ColumnView to C++.
This is a good idea for a host of reasons:

 - As long as C++ code can get a non-const reference to a Catalog,
   Python code can't possibly invalidate that cache reliably, because
   it can't tell when the Catalog is mutated.  We're basically
   assuming that only Python code modifies a Catalog that has a cached
   ColumnView, and (apparently) getting away with it, but that's at
   some level just luck, even though typical usage patterns _should_
   make it rare.

 - People don't expect to see a class's state split between C++ and
   Python (even a cache).  It's unusual and surprising.

 - Having this cache in Python forces the implementation of
   __getitem__ to straddle the C++/Python boundary in an awkward way,
   and that's getting in the way of making it go faster.

At the same time, I've made it so attempting to get the column view
for a noncontiguous catalog results in an empty std::optional return,
not an exception, *in C++*.  In Python, the behavior is the same for
backwards compatibility, but I'm pretty confident that this method was
never called in C++ (aside from those Python bindings), and if not,
we'll get an obvious compile failure downstream.  We also cache the
fact that a catalog is not contiguous, rather than recomputing it
each time.