Move most of Catalog.__getitem__ to C++.
Combined with the other changes on this ticket, quick benchmarks
show that this is:

 - 3.8-4.6x faster for by-name column array access (for views)
 - 1.2-1.5x faster for Key column array access (for views)
 - 2.6-4.3x faster for integer row indexing
 - about the same for slice and boolean-array indexing of catalogs.

This wasn't exhaustively checked - some Key types will be slower than
others, due to the way pybind11 overloads are dispatched, and the
non-type-dispatch parts of these calls sometimes scales with the number
of rows or columns.  But overall this speeds things up, in some cases
by a lot.

By-name column access is now a bit faster than Key access, which is
nice because names are much more natural in Python.  That's possible
because type dispatch in pure C++ with std::variant (probably a switch
or vtable under the hood) is much faster than pybind11 type-conversion
dispatch (a loop over all types until one matches).