Replace FITS I/O functions with overloads that don't use shared_ptr.
The original versions of these functions violate rule 5-24b of the
LSST C++ style guide and/or rule F7 of the C++ Core Guidelines: they
take a smart pointer argument but don't actually need to, because they
don't take ownership of the object or otherwise manipulate the
object's lifetime.  This is a viral problem, because it tempts
downstream code to use smart pointers unnecessarily in arguments that
will be forwarded to these.

There are many more examples of this, but they'll be easier to resolve
with or after RFC-840.