Reset the Fits object status to 0 if we are throwing an error
The nature of the cfitsio library is that many calls fail when the status value passed in is not set to a value indicating "success".

Prior to this patch we did not reset that status value after throwing an error, and since the status value is part of the state of the Fits class, any operation after an exception was thrown was liable (though not guaranteed) to fail purely because some previous (invalid) operation failed.

Most notably, when moving to an invalid HDU and back again, moving back would fail for this reason.

Another possible solution to this would be to manually reset status every time we make a cfitsio call, or to alter the LSST_FITS_EXCEPT macro to take a status allowing us to store a temporary and reset the class held value before throwing. The solution chosen, was to construct the exception as per previously, then store it, reset the class status and finally throw the exception.