Relocate exposure summary stats <-> afw.table code here.
...from lsst.pipe.tasks.ConsolidateVisitSummaryTask, which now no
longer depends on the details of what's in the summary stats.

The schema creation code is essentially unchanged; it duplicates all
field names and docs because I couldn't find a way to extract the
docstrings from the attributes that isn't worse than the duplication.

The update_record code is rewritten to use dataclasses.fields and
hence not depend on the fields in too much detail, and there's now
a from_record method for the opposite direction.