Trap for INHERIT having an undefined value
This should never happen in a well-formed FITS header (the standard
does not allow it) but in some LATISS test data it does have
an undefined value and this breaks the butler's attempts to
read the file.

This simple change checks for undefined and defaults to False
in that situation. This is not exactly what the LATISS data
need but does allow the file to be read.

Example file: AT_O_20220429_000008_R00_S00.fits