Stop supporting decompression into [u]int64 images.
CFITSIO very clearly (according to its error message when you try)
compressing [u]int64 images.  It also fails when you try to decompress
an image with a smaller integer pixel size into uint64, but with a
much more garbled error message, and only on macOS.  Easiest to just
not claim to support that.